feat(feign): openfeign配置可选依赖
