+++ wzkris-common/wzkris-common-openfeign/src/main/java/com/wzkris/common/openfeign/config/OpenFeignConfig.java
@@ -1,6 +1,7 @@
 package com.wzkris.common.openfeign.config;
 
 import com.wzkris.common.openfeign.core.TargeterDecorator;
+import com.wzkris.common.openfeign.handler.FeignRequestAuthenticationInterceptor;
 import com.wzkris.common.openfeign.handler.FeignRequestInterceptor;
 import com.wzkris.common.openfeign.handler.FeignResponseInterceptor;
 import feign.RequestInterceptor;
@@ -9,11 +10,13 @@ import okhttp3.OkHttpClient;
 import okhttp3.Protocol;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanPostProcessor;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.cloud.openfeign.Targeter;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 import org.springframework.context.annotation.Bean;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 
 import java.util.Arrays;
 import java.util.concurrent.TimeUnit;
@@ -75,6 +78,12 @@ public class OpenFeignConfig implements ApplicationContextAware, BeanPostProcess
     @Bean
     public RequestInterceptor requestInterceptor() {
         return new FeignRequestInterceptor();
+    }
+
+    @Bean
+    @ConditionalOnClass(UsernamePasswordAuthenticationToken.class)
+    public RequestInterceptor requestAuthenticationInterceptor() {
+        return new FeignRequestAuthenticationInterceptor();
     }
 
     @Override
