+++ wzkris-common/wzkris-common-openfeign/src/main/java/com/wzkris/common/openfeign/handler/FeignAuthenticationFilter.java
@@ -6,12 +6,14 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
 
+@ConditionalOnClass(UsernamePasswordAuthenticationToken.class)
 public class FeignAuthenticationFilter extends OncePerRequestFilter {
 
     @Override
